feat: finished refactor + optimized + coding style
