+++ shell.c
@@ -9,23 +9,6 @@
 #include "history.h"
 #include "line_edition.h"
 #include <locale.h>
-#include <stdint.h>
-#include <stdio.h>
-
-typedef struct {
-    envdata_t *env;
-    uint8_t status;
-    bool exiting;
-} ShellContext;
-
-typedef struct {
-    char input[4096];
-    uint16_t input_len;
-    uint16_t cursor_pos;
-    char read[4];
-    uint8_t read_len;
-    bool is_tty;
-} InputBuffer;
 
 uint8_t write_prompt(envdata_t *env)
 {
@@ -33,234 +16,6 @@ uint8_t write_prompt(envdata_t *env)
         env->hostname, env->cwd);
     fflush(stdout);
     return 1;
-}
-
-int8_t utf8_char_len(uint8_t byte)
-{
-    return (byte & 0x80) == 0x00 ? 1 :
-        (byte & 0xE0) == 0xC0 ? 2 :
-        (byte & 0xF0) == 0xE0 ? 3 :
-        (byte & 0xF8) == 0xF0 ? 4 : -1;
-}
-
-uint8_t previous_utf8_char_length(const char* input, uint16_t cursor_position)
-{
-    uint8_t len = 1;
-
-    for (uint8_t i = 0; i < 3 && (input[cursor_position] & 0xC0) == 0x80; ++i) {
-        len++;
-        if (cursor_position > 0)
-            cursor_position--;
-        else
-            break;
-    }
-    return len;
-}
-
-bool process_arrow_keys(InputBuffer *input_data)
-{
-    if (input_data->read[2] == 'C' && input_data->cursor_pos <
-        input_data->input_len) {
-        input_data->cursor_pos += utf8_char_len
-            (input_data->input[input_data->cursor_pos + 1]);
-        printf("\x1b[C");
-        return true;
-    }
-    if (input_data->read[2] == 'D' && input_data->cursor_pos > 0) {
-        uint8_t prev_len = previous_utf8_char_length(input_data->input,
-            input_data->cursor_pos);
-        if (prev_len > input_data->cursor_pos)
-            return true;
-        input_data->cursor_pos -= prev_len == 1 ? 1 : prev_len;
-        printf("\x1b[D");
-        return true;
-    }
-    return false;
-}
-
-bool process_delete_key(InputBuffer *input_data)
-{
-    if (input_data->read[2] == '3' && input_data->read[3] == '~' &&
-    input_data->cursor_pos < input_data->input_len) {
-        int8_t len = utf8_char_len(input_data->input[input_data->cursor_pos]);
-        memmove(input_data->input + input_data->cursor_pos,
-                input_data->input + input_data->cursor_pos + len,
-                input_data->input_len - input_data->cursor_pos);
-        for (uint8_t i = 0; i <= len; i++)
-            input_data->input[input_data->input_len - i] = '\0';
-        input_data->input_len -= len;
-        printf("\x1B[P");
-        int8_t c_len = utf8_char_len(
-            input_data->input[input_data->cursor_pos]);
-        if (c_len > 1)
-            input_data->cursor_pos += c_len - 1;
-        return true;
-    }
-    return false;
-}
-
-bool process_home_end_keys(InputBuffer *input_data)
-{
-    if (input_data->read[2] == 'H' || (input_data->read[2] == '1' &&
-    input_data->read[3] == '~')) {
-        while (input_data->cursor_pos > 0) {
-            uint8_t prev_len = previous_utf8_char_length(input_data->input,
-                input_data->cursor_pos);
-            input_data->cursor_pos -= prev_len == 1 ? 1 : prev_len;
-            printf("\x1b[D");
-        }
-        return true;
-    }
-    if (input_data->read[2] == 'F' || (input_data->read[2] == '4' &&
-    input_data->read[3] == '~')) {
-        while (input_data->cursor_pos < input_data->input_len) {
-            input_data->cursor_pos += utf8_char_len(
-                input_data->input[input_data->cursor_pos]);
-            printf("\x1b[C");
-        }
-        return true;
-    }
-    return false;
-}
-
-void process_backspace_key(InputBuffer *input_data)
-{
-    if (input_data->cursor_pos <= 0 || input_data->input_len <= 0 ||
-    input_data->cursor_pos > input_data->input_len)
-        return;
-    uint8_t len = previous_utf8_char_length(input_data->input,
-        input_data->cursor_pos);
-    if (len > input_data->cursor_pos || input_data->cursor_pos - len < 0)
-        return;
-    if (input_data->cursor_pos >= len)
-        len = previous_utf8_char_length(input_data->input,
-        input_data->cursor_pos - previous_utf8_char_length(input_data->input,
-                                                    input_data->cursor_pos));
-    if (input_data->cursor_pos < input_data->input_len)
-        memmove(input_data->input + input_data->cursor_pos - len,
-            input_data->input + input_data->cursor_pos,
-            (input_data->input_len - input_data->cursor_pos));
-    for (uint8_t i = 0; i <= len; i++)
-        input_data->input[input_data->input_len - i] = '\0';
-    input_data->input_len -= len;
-    input_data->cursor_pos -= len;
-    printf("\x1B[D\x1B[P");
-}
-
-void process_enter_key(ShellContext *context, InputBuffer *input_data)
-{
-    int8_t len = utf8_char_len(input_data->read[0]);
-    if (len == 1 && input_data->read_len > 1)
-        len = input_data->read_len;
-    memcpy(input_data->input + input_data->cursor_pos, input_data->read, len);
-    input_data->input_len += len;
-    input_data->cursor_pos += len;
-    operate_on_previous_command(input_data->input, context->env->history);
-    if (input_data->input[0] == '!') return;
-    add_line_to_history(context->env->history, input_data->input);
-    if (input_data->is_tty) printf("\n");
-    context->status = run_user_input(input_data->input, context->env,
-                        &context->exiting);
-    if (!input_data->is_tty) context->exiting = true;
-    if (context->exiting) return;
-    memset(input_data->input, 0, sizeof(input_data->input));
-    input_data->input_len = input_data->cursor_pos = 0;
-    write_prompt(context->env);
-}
-
-bool add_key_at_end(InputBuffer *input_data, int8_t new_char_len)
-{
-    if (input_data->cursor_pos >= input_data->input_len) {
-        memcpy(input_data->input + input_data->cursor_pos,
-            input_data->read, new_char_len);
-        input_data->input_len += new_char_len;
-        input_data->cursor_pos += new_char_len;
-
-        input_data->input[input_data->input_len] = '\0';
-        if (input_data->is_tty)
-            printf("%.*s", new_char_len, input_data->read);
-        return true;
-    }
-    return false;
-}
-
-bool insert_new_utf8_before_current_utf8(InputBuffer *input_data,
-    int8_t new_char_len, int8_t current_char_len)
-{
-    if (current_char_len > 1 && new_char_len > 1) {
-        memmove(input_data->input + input_data->cursor_pos + new_char_len - 1,
-            input_data->input + input_data->cursor_pos - (current_char_len - 1),
-            (input_data->input_len - input_data->cursor_pos) +
-                                                    (current_char_len - 1));
-
-        memcpy(
-        input_data->input + input_data->cursor_pos - (current_char_len - 1),
-        input_data->read, new_char_len);
-
-        printf("\x1b[K%s",
-        input_data->input + input_data->cursor_pos - (current_char_len - 1));
-
-        input_data->input_len += new_char_len;
-        input_data->cursor_pos += new_char_len;
-        return true;
-    }
-    return false;
-}
-
-void insert_new_char_before_current_char(InputBuffer *input_data,
-    int8_t new_char_len, int8_t current_char_len)
-{
-    uint8_t offset = current_char_len > 1 ? 0 : new_char_len;
-
-    memmove(input_data->input + input_data->cursor_pos + offset,
-    input_data->input + input_data->cursor_pos - (current_char_len - 1),
-    (input_data->input_len - input_data->cursor_pos) + (current_char_len - 1));
-
-    memcpy(input_data->input + input_data->cursor_pos - (current_char_len - 1),
-        input_data->read, new_char_len);
-
-    printf("\x1b[K%s",
-        input_data->input + input_data->cursor_pos - (current_char_len - 1));
-
-    input_data->input_len += new_char_len;
-    input_data->cursor_pos += new_char_len;
-}
-
-void compute_and_move_cursor(InputBuffer *input_data)
-{
-    uint16_t i;
-    i = input_data->input_len - 1;
-    while (i >= input_data->cursor_pos) {
-        int8_t c_len = utf8_char_len(input_data->input[i]);
-        if (c_len <= 0)
-            c_len = previous_utf8_char_length(input_data->input, i);
-        i -= c_len;
-        printf("\x1b[D");
-    }
-}
-
-void process_regular_key(InputBuffer *input_data)
-{
-    if (input_data->input_len + input_data->read_len >
-    sizeof(input_data->input))
-        return;
-    int8_t new_char_len = utf8_char_len(input_data->read[0]);
-    if (new_char_len == 1 && input_data->read_len > 1)
-        new_char_len = input_data->read_len;
-    if (add_key_at_end(input_data, new_char_len)) return;
-    int8_t current_char_len = utf8_char_len(
-        input_data->input[input_data->cursor_pos]);
-    if (current_char_len <= 0)
-        current_char_len = previous_utf8_char_length(input_data->input,
-                                        input_data->cursor_pos);
-    if (!insert_new_utf8_before_current_utf8(input_data, new_char_len,
-    current_char_len))
-        insert_new_char_before_current_char(input_data, new_char_len,
-                                                            current_char_len);
-
-    input_data->input[input_data->input_len] = '\0';
-
-    compute_and_move_cursor(input_data);
 }
 
 void process_key(ShellContext *context, InputBuffer *input_data)
